// src/bitacoras/dto/crear-bitacora.dto.ts
export class CrearBitacoraDto {    
    id_Bitacora: number;
    id_Acreditado: number;
    id_Propuesta: number;
    id_Especialista: number;
    id_TipoActividad: number;
    s_Descripcion: string;
    id_Documento: number;
    s_DireccionIP: string;
    f_Inicial: string;
    f_Final: string;
    swc_Activo: boolean;
}